fix(web): /case-studies/[slug] — metadata propres par slug + slugs au sitemap (Closes #3435)
Le [slug] était 'use client' sans generateMetadata → canonical /case-studies
fixe pour les 12 études de cas, absentes du sitemap. Split serveur/client :
generateMetadata (title/description/canonical/OG par slug) ; sitemap émet
désormais les 12 slugs.
